docs: add Lessons and instructions
