+++ docs/wksp/05-agent-framework-workshop/create-deployable-app/index.md
@@ -1,0 +1,10 @@
+# Turning the console application into a Deployable application
+
+In this lesson we go over the steps on how to create a deployable version of the
+semantic kernel console application by creating 3 main components:
+
+1. [Backend API (C#) using ASP .NET Core libraries](backend-api.md)
+1. [React JS Web application (with Node Proxy)](web-app.md)
+1. [AZD + Bicep templates for Infrastructure as Code deployment](azd-infra.md)
+
+ ![Chatbot Application Diagram](../images/architecture.png)
