+++ workshop-agent-framework/dotnet/Lessons/Directory.Build.props
@@ -1,0 +1,18 @@
+<Project>
+    <Import Project="$([MSBuild]::GetPathOfFileAbove(Directory.Build.props, $(MSBuildThisFileDirectory)..))" />
+
+    <PropertyGroup>
+        <OutputType>Exe</OutputType>
+	</PropertyGroup>
+
+    <ItemGroup>
+        <ProjectReference Include="..\..\Core.Utilities\Core.Utilities.csproj" />
+    </ItemGroup>
+
+    <ItemGroup>
+        <None Update="appsettings.json">
+            <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+        </None>
+    </ItemGroup>
+
+</Project>
