+++ workshop-agent-framework/dotnet/Solutions/Lesson5/Lesson5.csproj
@@ -9,6 +9,8 @@
   </ItemGroup>
 
   <ItemGroup>
+    <!-- <PackageReference Include="Microsoft.Agents.AI" /> Temporarily removed due to package conflicts -->
+    <PackageReference Include="Azure.Identity" />
     <PackageReference Include="Microsoft.Agents.AI" />
   </ItemGroup>
