+++ workshop-agent-framework/dotnet/Solutions/Lesson6/Lesson6.csproj
@@ -9,12 +9,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <!-- <PackageReference Include="Microsoft.Agents.AI" /> Temporarily removed due to package conflicts -->
-    <PackageReference Include="Azure.Identity" />
-    <PackageReference Include="Microsoft.Agents.AI" />
-  </ItemGroup>
-
-  <ItemGroup>
     <None Update="appsettings.json">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
