+++ workshop-agent-framework/dotnet/dotnet.sln
@@ -20,6 +20,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Lesson6", "Solutions\Lesson
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Lesson7", "Solutions\Lesson7\Lesson7.csproj", "{A3E628A1-7D59-44A0-894D-C1FF2D8004A9}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Lesson8", "Solutions\Lesson8\Lesson8.csproj", "{B4E628A1-7D59-44A0-894D-C1FF2D8004B0}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -58,11 +60,16 @@ Global
 		{A3E628A1-7D59-44A0-894D-C1FF2D8004A9}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A3E628A1-7D59-44A0-894D-C1FF2D8004A9}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A3E628A1-7D59-44A0-894D-C1FF2D8004A9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B4E628A1-7D59-44A0-894D-C1FF2D8004B0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B4E628A1-7D59-44A0-894D-C1FF2D8004B0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B4E628A1-7D59-44A0-894D-C1FF2D8004B0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B4E628A1-7D59-44A0-894D-C1FF2D8004B0}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{26795795-0533-4764-9B02-8C75B19AF502} = {6258CA5D-F2BE-49EF-B2B4-06EC5EECAA0B}
 		{89561A59-D8BB-434D-A508-A69B18B52741} = {6258CA5D-F2BE-49EF-B2B4-06EC5EECAA0B}
 		{F2E628A1-7D59-44A0-894D-C1FF2D8004A8} = {6258CA5D-F2BE-49EF-B2B4-06EC5EECAA0B}
 		{A3E628A1-7D59-44A0-894D-C1FF2D8004A9} = {6258CA5D-F2BE-49EF-B2B4-06EC5EECAA0B}
+		{B4E628A1-7D59-44A0-894D-C1FF2D8004B0} = {6258CA5D-F2BE-49EF-B2B4-06EC5EECAA0B}
 	EndGlobalSection
 EndGlobal
